Update revisao Java OO

+++ exemplosDeCodigoJava/src/javaBasico/TesteTres.java
@@ -10,7 +10,7 @@ public class TesteTres {
         System.out.println("Digite o seu nome: ");
         String nome = scan.nextLine();
         System.out.println("Digite o seu CPF: ");
-        int cpf = scan.nextInt();
+        long cpf = scan.nextLong();
 
         System.out.println("Seu nome é :" + nome);
         System.out.println("Seu cpf é:" + cpf);
